Add docs and refactor some stuff
